Updates to make compatible with Mongodb 4.0
When used with mongoid 2.5 `writeConcern` options are being mapped
to `writeConcern` then also being included in the selector which is
causing problems with versions of mongo newer than 2.6.

`validate` is remove when appropriate. This is a little more slective
and conditional on the command.

Mapping `$updateAll` to `$update` with `$each` operation.